ramijraj-develop-ethical-drm-kit-for-creators/requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
ffmpeg-python>=0.2.0
imagehash>=4.3.1
flask>=2.3.0
requests>=2.31.0
telethon>=1.28.0
psutil>=5.9.0
cryptography>=41.0.0
pymongo>=4.5.0
sqlite3
pytest>=7.4.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
schedule>=1.2.0
python-magic>=0.4.27
tqdm>=4.65.0
aiohttp>=3.8.0